\name{annual.data}
\alias{annual.data}
\title{create an inter-annual dataset}
\description{create an inter-annual dataset by averaging the measured intra-annual isotope values}
\usage{annual.data(y, climate.iso)}
\arguments{
  \item{y}{study period over which to create an inter-annual dataset}
  \item{climate.iso}{time resolved intra-annual isotope measurements with matched climate data}
}
\value{
\item{annual.dataset}{inter-annual dataset}
}
\author{Alex Soudant
Maintainer: Alex Soudant <asoudant@hotmail.fr>}
\keyword{ programming }
\keyword{ isotope }
