\name{atm.correction}
\alias{atm.correction}
\title{
 create atmospheric correction values
}
\description{
 create atmospheric correction values from McCarroll and loader (2004) 
for years 1950 to 2003 and then extrapolate these values for years 2004 to 2010
}
\usage{
atm.correction()
}
\details{These values can be changed by user if another correction procedure is necessary/wanted}
\value{
%%  \item{atm.cor}{list of correction values to apply to carbon isotope time series}
}
\references{McCarroll and loader (2004)}
\author{Alex Soudant
Maintainer: Alex Soudant <asoudant@hotmail.fr>}
\keyword{ programming }
\keyword{ correction }
