\name{cell.lifetime}
\alias{cell.lifetime}
\title{
set the period of time for cell maturation phases
}
\description{
set the period of time for cell maturation phases in days of the year
}
\usage{
cell.lifetime()
}
\value{
\item{increment.division}{days of division phase}
\item{increment.enlargment}{days of enlargment phase}
\item{increment.wallthick}{days of wall thicknening phase}
\item{increment.total}{days for total maturation}
}
\author{
Alex Soudant
Maintainer: Alex Soudant <asoudant@hotmail.fr>
}
\keyword{ cell }
\keyword{ programming }
