\name{intraannual.data}
\alias{intraannual.data}
\title{produce the intra-annual dataset}
\description{produce the intra-annual dataset with matched time resolved isotopes and climate data}
\usage{
intraannual.data(y, final.isotope)
}
\arguments{
  \item{y}{study period over which to produce the intra-annual dataset}
  \item{final.isotope}{time resolved isotope data}
}
\value{
\item{climate.iso}{dataset with matched time resolved isotopes and climate data}
}
\author{Alex Soudant
Maintainer: Alex Soudant <asoudant@hotmail.fr>}
\keyword{ programming }
\keyword{ dataset }
