\name{isoclim}
\alias{isoclim}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{construct a work environment, cast programs for data manipulation, destruct work environment.}
\description{ Run all other functions in the intended logical order for data manipulation}
\usage{
isoclim()
}
\details{running isoclim will not display derived datasets and parameters to fit the Gompertz}
\author{Alex Soudant
Maintainer: Alex Soudant <asoudant@hotmail.fr>}
\keyword{ programming }

