\name{isoclim_dem-package}
\alias{isoclim_dem-package}
\alias{isoclim_dem}
\docType{package}
\title{Climate Influence Analysis over Stable Isotope Time Series in R}
\description{This package takes vegetation stable isotopes and 
flux-tower climate data to estimate the growing season, fit Gompertz
curves accordingly to the life properties of wood cells and fit 
linear regressions between intra-annually time resolved isotopes 
and co-occuring climate data}
\details{
\tabular{ll}{
Package: \tab isoclim\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-06-07\cr
License: \tab GPL (>= 2)\cr
}
}
\author{Alex Soudant
Maintainer: Alex Soudant <asoudant@hotmail.fr>
}
\references{Soudant et al. (2015)}
\keyword{ package }

