\name{match.climate}
\alias{match.climate}
\title{match the time resolved isotopes with Flux-tower data}
\description{match the time resolved isotopes with Flux-tower data year by year
with the matching period depending on which cell phases selected}
\usage{
match.climate(y, cellphase, final.isotope)
}
\arguments{
  \item{y}{study period over which to match the time resolved isotopes with Flux-tower data}
  \item{cellphase}{which cell phases to match isotopes with climate data}
  \item{final.isotope}{time resolved isotope time series}
}
\author{Alex Soudant
Maintainer: Alex Soudant <asoudant@hotmail.fr>}
\keyword{ programming}
\keyword{ isotope}
\keyword{ cellphase}