\name{plot.correlation}
\alias{plot.correlation}
\title{
plot correlation figures for climate variables month by month
}
\description{
Single linear correlation applied between inter-annual d13C values and climate variables month by month and displayed in a graphic  
}
\usage{
plot.correlation(annual.dataset, climate.iso)
}
\arguments{
  \item{annual.dataset}{
annual dataset from carbon isotope values
}
  \item{climate.iso}{
time resolved and corrected intra-annual carbon isotope dataset
}
}
\author{
Alex Soudant
Maintainer: Alex Soudant <asoudant@hotmail.fr>
}
\keyword{ correlation }
\keyword{ programming }
