\name{plot.interannual}
\alias{plot.interannual}
\title{set the linear regressions between inter-annual stable isotopes and climate}
\description{fit and plot the linear regressions between inter-annual stable isotopes and three climate variables}
\usage{
plot.interannual(y, annual.dataset)
}
\arguments{
  \item{y}{study period over which to fit and plot the linear regressions}
  \item{annual.dataset}{inter-annual and time resolved isotope time series}
}
\author{Alex Soudant
Maintainer: Alex Soudant <asoudant@hotmail.fr>}
\keyword{programming}
\keyword{isotope}
\keyword{climate}