\name{plot.rawseries}
\alias{plot.rawseries}
\title{plot the raw series of stable isotopes}
\description{plot three raw series of stable isotopes that the user has implemented in the data folder}
\usage{
plot.rawseries(iso1, iso2, iso3, sample.res)
}
\details{
do not need argument "y" as each isotope time series will be plotted against the full period present in the dataset
}
\arguments{
  \item{iso1}{first time series of stable isotopes}
  \item{iso2}{second time series of stable isotopes}
  \item{iso3}{third time series of stable isotopes}
  \item{sample.res}{resolution at which the intra-annual stable isotope vales have been sampled in microns}
}
\author{Alex Soudant
Maintainer: Alex Soudant <asoudant@hotmail.fr>}
\keyword{programming}
\keyword{isotope}
