\name{plot.standardseries}
\alias{plot.standardseries}
\title{plot the procedure to standardise time series of isotope data}
\description{
calculates and plot atmopheric corrected time series of isotope data according to McCarroll and Loader (2004)
}
\usage{
plot.standardseries(y, isotope.data, atm.cor)
}
\arguments{
  \item{y}{
study period over which to calculate the atmopheric corrected time series of isotope data
}
  \item{isotope.data}{
standardised time series of isotope data
}
  \item{atm.cor}{
atmopheric correction values
}
}
\value{
\item{cor.iso}{atmopheric corrected time series of isotope data}
}
\references{McCarroll and Loader (2004)}
\author{Alex Soudant
Maintainer: Alex Soudant <asoudant@hotmail.fr>}
\keyword{programming}
\keyword{isotope}

