\name{rad.pot}
\alias{rad.pot}
\title{calculation of potential downwelling shortwave radiation}
\description{
Calculates the potential downwelling shortwave radiation at the surface
according to Stull(1988), 7.3.1 (p.257) assuming no cloud cover
}
\usage{
rad.pot(y, latitude, longitude)
}

\arguments{
  \item{y}{study period over which to calculate the potential downwelling shortwave radiation}
  \item{latitude}{latitude of the study site}
  \item{longitude}{longitude of the study site}
}
\references{Stull(1988), 7.3.1 (p.257)}
\author{Alex Soudant
Maintainer: Alex Soudant <asoudant@hotmail.fr>}
\keyword{programming}
\keyword{radiation}

